# protocol/http/v1/openapi.yaml
...
openapi: 3.1.0
info:
  title: FileBelt HTTP API
  version: 1.0.0
  description: Durable FileBelt control-plane, storage-grant, collaboration, and outbound MCP broker contract.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/auth/callback:
    get:
      operationId: completeOidcLogin
      security: []
      parameters:
        - name: code
          in: query
          schema:
            type: string
            maxLength: 4096
        - name: state
          in: query
          schema:
            type: string
            maxLength: 4096
        - name: error
          in: query
          schema:
            type: string
      responses:
        "303":
          description: Establish an opaque local session and redirect to the allowlisted local path.
          headers:
            Set-Cookie:
              schema:
                type: string
            X-FileBelt-Csrf:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: filebelt_session
    capabilityHeader:
      type: http
      scheme: fbcap1
      description: "Send `Authorization: fbcap1 <base64url-envelope>`."
    capabilityCookie:
      type: apiKey
      in: cookie
      name: filebelt_capability
...
  responses:
    Problem:
      description: Stable FileBelt problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          maxLength: 128
          pattern: "^[a-z][a-z0-9]*(?:\\.[a-z0-9_-]+)+$"
          description: Stable dotted FileBelt reason code. MCP failures use the `mcp.*` namespace.
...